refactor: #148 フォローモーダル削除
